2657365 IPFilter 4.1.32 RC4 fails to compile on FreeBSD7/AMD64

+++ ipfilter/ip_fil.h
@@ -627,7 +627,7 @@ typedef	struct	frentry {
 #define	FR_NOLOGTAG	0
 
 #ifndef	offsetof
-#define	offsetof(t,m)	(int)((&((t *)0L)->m))
+#define	offsetof(t,m)	(size_t)((&((t *)0L)->m))
 #endif
 #define	FR_CMPSIZ	(sizeof(struct frentry) - \
 			 offsetof(struct frentry, fr_func))
