+++ ipfilter/tools/ippool.c
@@ -497,7 +497,7 @@ char *argv[];
 		}
 
 	}
-	printf("%u object%s flushed\n", flush.iplf_count,
+	printf("%d object%s flushed\n", (int) flush.iplf_count,
 	       (flush.iplf_count == 1) ? "" : "s");
 
 	return 0;
